categorize windows.psxview as malware plugin
